# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/warga": {
            "get": {
                "tags": [
                    "Warga"
                ],
                "summary": "Get all warga data",
                "description": "Returns all warga data",
                "operationId": "getAllWarga",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "an 'unexpected error' occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Warga"
                ],
                "summary": "Store new warga",
                "description": "Stores new warga data",
                "operationId": "storeWarga",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Warga"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Warga data stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Data validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "an 'unexpected error' occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/warga/{id}": {
            "get": {
                "tags": [
                    "Warga"
                ],
                "summary": "Get warga data by id",
                "description": "Returns warga data based on given id",
                "operationId": "getWargaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the warga to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Warga not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "an 'unexpected error' occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Warga"
                ],
                "summary": "Update existing warga data",
                "description": "Updates existing warga data based on given id",
                "operationId": "updateWarga",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the warga to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Warga"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Warga data updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DataResponse": {
                "description": "Data response dari permintaan perizinan",
                "properties": {
                    "status": {
                        "description": "Keterangan status response",
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "description": "Pesan terkait response",
                        "type": "string",
                        "example": "Permintaan perizinan berhasil diajukan"
                    },
                    "data": {
                        "description": "Data API",
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "status": "success",
                    "message": "Permintaan perizinan berhasil diajukan",
                    "data": {
                        "nama": "John Doe",
                        "alamat": "Jl. Kenanga No. 123",
                        "jenis_perizinan": "Izin Usaha",
                        "tanggal_mulai": "2023-04-01",
                        "tanggal_selesai": "2023-06-30"
                    }
                }
            },
            "Error": {
                "description": "Error",
                "properties": {
                    "message": {
                        "description": "Pesan error",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Oops, something went wrong."
                }
            },
...
            "NotFound": {
                "description": "Data yang diminta tidak ditemukan",
                "properties": {
                    "message": {
                        "description": "Pesan error yang menunjukkan data tidak ditemukan",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Data dengan ID 123 tidak ditemukan"
                }
            },
...
            "Warga": {
                "properties": {
                    "warga_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nama_warga": {
                        "type": "string"
                    },
                    "alamat": {
                        "type": "string"
                    },
                    "nomor_telepon": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "nik": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}